backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
nltk
youtube_search
ultralytics
python-multipart
Pillow
ollama
